Fix quoting; everything after '>' is a quote
Plenty of people voiced their opinion that everything after '>' should be a
quote. Trying to conform to anything other brings only chaos and makes
situation worse, since it is impossible to catch every exception.

There are already plenty of cases where regexp simply can't work with unicode.

In the future, qTox should have easy to use r-click options for 'Paste text as
 a quote' and 'Copy as a quote', which makes it a must to have quote detection
as simple as possible, to avoid situation where user wants to paste something
as a quote, and suddenly it doesn't seem to work.

Fixes #1059 #1099 #1173
Continuation of #1332