Possible Fix for Tray Icon Detection
Call it a hunch, but XDG_SESSION_DESKTOP and XDG_CURRENT_DESKTOP don't seem to be as interchangeable as they seem to be. Notably, the XDG_CURRENT_DESKTOP environment variable in Ubuntu GNOME 15.04 (systemd woo) is set like this:

XDG_SESSION_DESKTOP=gnome

which is a variable qtox is expecting from XDG_CURRENT_DESKTOP. On the other hand:

XDG_CURRENT_DESKTOP=GNOME

The casing is different, obviously. Two solutions emerge: change the casing in those if statements, or change the environment variable we're checking. I've done the later.

Even if this solution doesn't work at all, my guess as to why the tray icons for all this time is this environment variable setup. Also to be noted is that XDG_SESSION_DESKTOP takes one item, whereas XDG_CURRENT_DESKTOP takes a colon-seperated list.

Some stuff about XDG_SESSION_DESKTOP is over here:

http://www.freedesktop.org/software/systemd/man/pam_systemd.html