fix(capslockindicator): fix altering the line edit height
Fix altering the Line Edit widget height when the indicator is on (e.g.
when caps lock is active).

To avoid messing with the style and all the complex issues that arise
from it, use an action to display the icon inside the line edit, as it
is a built-in feature that work just fine.

Fixes #3379.