refactor: replace CapsLockIndicator with new PasswordEdit widget
Introduce a new PasswordEdit widget extending QLineEdit that takes care
of all the specifics of a QLineEntry when it is used to input a
password, including echo mode and caps lock indicator.

Also optimize the event handling to only listen to global events when
it is actually needed, e.g. when a password field is actually visible.