feat(widget): add support for QtKeychain to store/retrieve passwords
This commit adds support for QtKeychain, a platform-independent Qt API
for storing passwords securely. According to its documentation, it
stores passwords depending on operating system and desktop environment:
- Mac OS X's Keychain
- Windows' Credential Store
- Linux/KDE: KWallet
- Linux/Gnome: Keyring

Using this library is optional and can be enabled/disabled at compile
time: The feature is enabled by default, but can be disabled by
passing 'DISABLE_QTKEYCHAIN=YES' to a qmake call.

Even if qTox is compiled against QtKeychain, the user can enable/
disable storing the password in the login screen (global setting, not
per profile).
Later, in the privacy settings, the user can choose to clear the
current profile's stored password.