chore: Statically link the windows platform plugin.
When building statically linked versions of qTox, we need to link in all
plugins it uses. Linking the windows platform plugin will make linking
succeed, and the program will start. Some additional plugins may be
needed for e.g. image formats. These require some testing on windows and
are out of scope for this PR.