fix(UI): fix own status message not being properly displayed in tooltip
This causes a slight regression where own status can become quite huge
when user sets e.g. 120 newlines between first and last character of the
status message.

Given that this fixes a problem with incorrect (broken) formatting of
display of own status messages, the slight regression is insignificant.

There is also a Qt bug involved that causes formatting to be broken if
text contains "HTML" tags.
https://bugreports.qt.io/browse/QTBUG-57477

Used workaround requires using additional HTML to preserve newlines even
if HTML is a part of the status message.

Related to #935.