refactor: onTextEditChanged method refactoring
Made as single commit because this change requires an explanation. As i
understand, `getTypingNotification` always return the same value because
`setTypingNotification` is not connected for any signal and the only
time it's called - from `PrivacyForm`'s method which is never called and
also is not connected to signals. I guess that removing `getTypingNotification`
from `onTextEditChanged` will not affect anything