refactor(windows): keep build script rerunnable
The build script should not fail if you re-run it from inside the Docker
container it has already ran in. This makes it easier to patch and test
build errors, since you can get into a Docker container, run the build
script, observe the failure, patch the build script and re-run it again.