fix(group): don't display netcam view for group voice calls
This is a minimal fix to reduce risk for the release. A more complete
re-architecture will be made.

The netcam covers much of the chat in groups, and has nothing to show since
group video calls aren't possible. Who is speaking in call is already shown by
the bold names at the top of the group, taking much less space.

Fix #5918