feat(chat): Implement global hotkey capture and push-to-talk
The mute option in each call is toggled when the key combo is pressed, so this
feature is also effectively push-to-mute as well if users are unmuted in chat
to start.

Load libuiohook at runtime because it registers with the OS on library load,
and segfaults if no window system is available, e.g. for unit tests.