fix(settings): Make all Settings connections queued to avoid deadlocks
When a setting is set it locks the Settings mutex and emits a signal for that
setting. If a slot is connected to that signal and lives on the same thread as
Settings, it is executed immediately with the Settings mutex still locked. That
slot can then lock a series of other mutexes. During this time another thread
can lock a mutex and then try to read settings, which will cause a deadlock due
to the opposite order of multiple mutex locking. By making all Settings
connectiong queued, the Settings setter will return and unlock the Settings
mutex, ensuring that getters won't ever be blocked indefinitely.

Fix #6218