refactor: Provide a Virtual Method Anchor for Classes in Headers
Define at least one virtual method in cpp files for classes with
vtables to improve link time. This is done by defining a defaulted
constructor in the cpp file, or moving an overridden method from the
header to the cpp file. Where neccessary, create new cpp files and
add them to CMakeLists.

See http://llvm.org/docs/CodingStandards.html#provide-a-virtual-method-anchor-for-classes-in-headers.

Fixes #6036.