chore(CI): Don't try to upload installer on debug builds
The installer is only built on release builds. We used to ignore failure to
upload prior to 5fcf86b5212b1c36686eaacf9c844ceb0e15adb1. We stopped ignoring
because a07ab89cc84bcab3a75277dd4efd1608535127bb tried to make our upload
targetted, but it only resolved the issue for the zip, not the installer.