chore(CI): Handle build scripts called with relative paths with spaces
foo=dirname "$0" only works if the script is called with an absolute path and
that path contains no spaces. Add quotes everywhere to contain spaces in
directory names or file names, and add realpath to get the absolute path
regardless of call path.

Not necessarily needed, but like in e40586803743c9f08a9018caccf8038f754a31e4
it can avoid some annoying issues when e.g. that script was called with a
relative path in CI.