feat(history): Remove group history on quit when selected
The group itself is a peer, as well as all the members of the group. Members of the group have associated aliases, but the group does not. Chat messages reference the chat_id of the group, with senders referencing the member peer.

Deleting all history belonging to the chat_id already generically cleans up group history. Users who only have history from that group need to have their aliases and peer entries removed.

We don't store what users or aliases where in which group, so just remove any alias that has no associated chat message with it, and and peer that doesn't have an alias (friend with history) and doesn't have a chat message associated with it as a chat_id (group).

The group itself is removed from the peer table directly, and the group never has an alias entry.