fix(history): Don't cast file_id to QString before inserting
The implicit cast in addNewFileMessage's argument causes a conversion from the
binary data of file_id to be interpreted as UTF8 characters. On null bytes or
invalid UTF8, the resulting QString can be empty or truncated. Empty IDs causes
null file_restart_id's to be inserted into the database.

Fixes the cause of #6553, but the database damage still needs to be healed.